테스트 클래스 분리
UserServiceTest 클래스 생성, 유저 관련 테스트 로직 옮김
