+++ src/test/java/RequestTest.java
@@ -1,14 +1,10 @@
 import exception.HttpErrorMessage;
 import exception.HttpException;
-import exception.UserErrorMessage;
-import model.User;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import util.RequestParser;
-import webserver.UserService;
 
 import java.io.UnsupportedEncodingException;
-import java.util.HashMap;
 import java.util.Map;
 
 import static org.assertj.core.api.Assertions.*;
@@ -50,35 +46,5 @@ public class RequestTest {
                     Map.entry("password", "password"),
                     Map.entry("name", "박재성"),
                     Map.entry("email", "javajigi@slipp.net")));
-    }
-
-    @Test
-    @DisplayName("유저 생성 테스트")
-    void createUser() {
-        UserService userService = new UserService();
-        User user = userService.addUser(Map.ofEntries(
-                Map.entry("userId", "javajigi"),
-                Map.entry("password", "password"),
-                Map.entry("name", "박재성"),
-                Map.entry("email", "javajigi@slipp.net")));
-        assertThat(userService.getUsers().contains(user));
-    }
-
-    @Test
-    @DisplayName("중복된 유저 생성 안 되어야 함")
-    void createDuplicateUser() {
-        UserService userService = new UserService();
-        userService.addUser(Map.ofEntries(
-                Map.entry("userId", "javajigi"),
-                Map.entry("password", "password"),
-                Map.entry("name", "박재성"),
-                Map.entry("email", "javajigi@slipp.net")));
-        assertThatThrownBy(() -> {
-            userService.addUser(Map.ofEntries(
-                    Map.entry("userId", "javajigi"),
-                    Map.entry("password", "password"),
-                    Map.entry("name", "박재성"),
-                    Map.entry("email", "javajigi@slipp.net")));
-        }).isInstanceOf(RuntimeException.class).hasMessageContaining(UserErrorMessage.DUPLICATE_USER.getMessage());
     }
 }
